Security and settings update
