+++ frontend/src/App.tsx
@@ -14,6 +14,7 @@ const StudentCourses = lazy(() => import('./pages/StudentCourses'))
 const InstructorCourses = lazy(() => import('./pages/InstructorCourses'))
 const HeadCourses = lazy(() => import('./pages/HeadCourses'))
 const Settings = lazy(() => import('./pages/Settings'))
+const Safety = lazy(() => import('./pages/Safety'))
 
 // Loading fallback component for route transitions
 function RouteLoadingFallback() {
@@ -67,6 +68,9 @@ function App() {
       {/* Legacy routes - redirect to role-specific routes */}
       <Route path="/settings" element={<Layout />}>
         <Route index element={<Settings />} />
+      </Route>
+      <Route path="/security" element={<Layout />}>
+        <Route index element={<Safety />} />
       </Route>
       <Route path="/lecturer" element={<Navigate to="/instructor" replace />} />
       <Route path="/course/:id" element={<Navigate to="/" replace />} />
